@@ -1,16 +1,17 @@
 {
   "name": "beantown",
-  "version": "0.1.45",
+  "version": "0.1.46",
   "private": true,
   "type": "module",
   "engines": {
-    "node": ">=14.17.6 <18"
+    "node": "<18"
   },
   "scripts": {
-    "start": "node --trace-warnings --unhandled-rejections=strict ./bin/server",
+    "start": "node --trace-warnings --unhandled-rejections=strict ./bin/server.js",
     "build": "npx webpack --config webpack.config.js"
   },
   "dependencies": {
+    "@aws-sdk/client-secrets-manager": "^3.398.0",
     "@babel/runtime": "^7.17.8",
     "axios": "^0.24.0",
     "compression": "^1.7.4",
