@@ -1,26 +1,36 @@
-var createError = require('http-errors')
-const bodyParser = require('body-parser')
-var compression = require('compression')
-var express = require('express')
-var path = require('path')
-var cookieParser = require('cookie-parser')
-var logger = require('morgan')
+import HttpError from 'http-errors'
+import bodyParser from 'body-parser'
+import compression from 'compression'
+import express from 'express'
+import path from 'path'
+import * as uuid from 'uuid'
+// var cookieParser = require('cookie-parser')
+import cookieParser from 'cookie-parser'
+import morgan from 'morgan'
 // var sassMiddleware = require('node-sass-middleware')
-var session = require('express-session')
-var indexRouter = require('./routes/index.js')
-var merchRouter = require('./routes/merch')
-var contactRouter = require('./routes/contact')
-var menuRouter = require('./routes/menu')
-const redis = require('redis')
-const redisStore = require('connect-redis')(session)
-var app = express()
-
+// var session = require('express-session')
+import session from 'express-session'
+import * as indexRouter from './routes/index.js'
+import * as contactRouter from './routes/contact.js'
+import * as menuRouter from './routes/menu.js'
+import * as merchRouter from './routes/merch.js'
+import redis from 'redis'
+import redisStore from 'connect-redis'
+import { fileURLToPath } from 'url'
+const redisSession = redisStore(session)
+const app = express()
+console.log(`UUID: ${Object.keys(uuid)}`)
+console.log(`SESSION: ${Object.keys(session)}`)
+console.log(`REDIS SESSION: ${Object.keys(redisSession)}`)
+console.log(`REDIS STORE: ${Object.keys(redisStore)}`)
 let redisClient = redis.createClient({ legacyMode: true })
 redisClient.connect().catch(console.error)
 
-const { v4: uuidv4 } = require('uuid')
-const logLevel = process.env.LOG_LEVEL || 'tiny'
-
+const { v4: uuidv4 } = uuid.v4
+const logLevel = process.env.LOG_LEVEL || 'dev'
+console.log(logLevel)
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 // view engine setup
 app.set('views', path.join(__dirname, 'views'))
 app.set('view engine', 'ejs')
@@ -29,9 +39,9 @@ app.use(compression())
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: false }))
 if (process.env.NODE_ENV === "development") {
-  app.use(logger(logLevel))
+  app.use(morgan(logLevel))
 } else {
-  app.use(logger(logLevel, {
+  app.use(morgan(logLevel, {
     skip: function (req, res) { return res.statusCode < 400 }
   }))
 }
@@ -53,22 +63,22 @@ app.use(session({
   saveUninitialized: true,
   cookie: { maxAge: 6000000, path: '/merch/cart' },
   name: "cartId",
-  store: new redisStore({ host: 'localhost', port: 6379, client: redisClient, ttl: 86400 }),
+  store: new redisSession({ host: 'localhost', port: 6379, client: redisClient, ttl: 86400 }),
 }))
 app.use(express.static(path.join(__dirname, '../dist/public')))
 // app.use(express.static(staticUrl));
 
 const urlRoot = process.env.NODE_JAL_URL_ROOT || '/'
 
-app.use(urlRoot, indexRouter)
-app.use(urlRoot + 'merch', merchRouter)
-app.use(urlRoot + 'contact', contactRouter)
-app.use(urlRoot + 'menu', menuRouter)
+app.use(urlRoot, indexRouter.default)
+app.use(urlRoot + 'merch', merchRouter.default)
+app.use(urlRoot + 'contact', contactRouter.default)
+app.use(urlRoot + 'menu', menuRouter.default)
 // app.use(urlRoot + 'cart', cartRouter)
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
-  next(createError(404))
+  next(HttpError(404))
 })
 
 // error handler
@@ -82,4 +92,4 @@ app.use(function(err, req, res, next) {
   res.redirect('/error')
 })
 
-module.exports = app
+export default app
