@@ -1,48 +1,48 @@
 import React from 'react'
 import {
-    StyledAnchor,
-    StyledInfo,
-    StyledLinkButton } from './styles.js'
+  StyledAnchor,
+  StyledInfo,
+  StyledLinkButton } from './styles.js'
 
 
 export const InfoSection = (props) => {
-    return (
-        <StyledInfo
-            color={props.bgColor}
-            fontColor={props.fontColor}
-            id={`${props.slug}-div`}
-            textAlign={props.textAlign}
-            marginTop={props.marginTop}
-            paddingTop={props.paddingTop}
-            paddingBottom={props.paddingBottom}
-            paddingLeft={props.paddingLeft}
-            paddingRight={props.paddingRight}>{props.children}</StyledInfo>
-    )
+  return (
+    <StyledInfo
+      color={props.bgColor}
+      fontColor={props.fontColor}
+      id={`${props.slug}-div`}
+      textAlign={props.textAlign}
+      marginTop={props.marginTop}
+      paddingTop={props.paddingTop}
+      paddingBottom={props.paddingBottom}
+      paddingLeft={props.paddingLeft}
+      paddingRight={props.paddingRight}>{props.children}</StyledInfo>
+  )
 }
 
 export const Title = (props) => {
-    return (
-        <h1>{props.children}</h1>
-    )
+  return (
+    <h1>{props.children}</h1>
+  )
 }
 
 export const LinkButton = (props) => {
-    const goTo = () => {
-        window.location.href=props.url
-    }
-    return (
-        <StyledLinkButton buttonColor={props.color} fontColor={props.fontColor}>
-            <button onClick={goTo} type="button">{props.children}</button>
-        </StyledLinkButton>
-    )
+  const goTo = () => {
+    window.location.href=props.url
+  }
+  return (
+    <StyledLinkButton buttonColor={props.color} fontColor={props.fontColor}>
+      <button onClick={goTo} type="button">{props.children}</button>
+    </StyledLinkButton>
+  )
 }
 
 export const Anchor = (props) => {
-    return (
-        <StyledAnchor>
-            <a href={props.target}>{props.text}</a>
-        </StyledAnchor>
-    )
+  return (
+    <StyledAnchor>
+      <a href={props.target}>{props.text}</a>
+    </StyledAnchor>
+  )
 }
 
 
