@@ -1,1 +1,1 @@
-export * from "./icons"
+export * from './icons.js'
