@@ -5,38 +5,38 @@ import { ImageSlider } from '../imageSliders/main.js'
 import { config } from '../../utils/main.js'
 import { sendEvent } from '../kafka/index.js'
 const COLORS = config.colors
-
+const STATIC_PATH = `${config.urls.static}/img/merch`
 
 const images = [
-    "bt_pool_tables.jpeg",
-    "bt_private_event_1.jpeg",
-    "bt_private_event_pt2.jpeg",
-    "bt_private_event_pt4.jpeg",
-    "bt_private_event_2.jpeg",
-    "elisha_brown_room.jpeg",
-    "elisha_brown_room2.jpeg"
+  "bt_pool_tables.jpeg",
+  "bt_private_event_1.jpeg",
+  "bt_private_event_pt2.jpeg",
+  "bt_private_event_pt4.jpeg",
+  "bt_private_event_2.jpeg",
+  "elisha_brown_room.jpeg",
+  "elisha_brown_room2.jpeg"
 ]
 
 const sliderStyles = {
-    marginTop: "1rem",
-    marginRight: "unset",
-    marginLeft: "unset",
-    width: "95vw",
-    img: {}
+  marginTop: "1rem",
+  marginRight: "unset",
+  marginLeft: "unset",
+  width: "95vw",
+  img: {}
 }
 
 const sliderSettings = {
-    dots: true,
-    autoplay: true,
-    speed: 1000,
-    autoplaySpeed: 8500,
-    pauseOnHover: false,
-    swipeToSlide: true,
-    fade: true,
-    infinite: true,
-    arrows: true,
-    slidesToShow: 1,
-    slidesToScroll: 1
+  dots: true,
+  autoplay: true,
+  speed: 1000,
+  autoplaySpeed: 8500,
+  pauseOnHover: false,
+  swipeToSlide: true,
+  fade: true,
+  infinite: true,
+  arrows: true,
+  slidesToShow: 1,
+  slidesToScroll: 1
 }
 
 
@@ -46,7 +46,11 @@ export const PrivateParties = () => {
   })
   const imagePath = "https://static.prod.beantownpub.com/img/slider"
   return (
-    <StyledPartiesContainer id="privatePartiesContainer" backgroundColor={COLORS.yellow} aria-details="Parties container">
+    <StyledPartiesContainer
+      id="privatePartiesContainer"
+      backgroundColor={COLORS.yellow}
+      aria-details="Parties container"
+    >
       <h1>Private Parties</h1>
       <article id="privatePartiesArticle">
       Make your next private event a memorable one. Beantown Pub offers a variety of spaces that are perfect for accomadating small gatherings
