@@ -26,21 +26,21 @@ export const StyledPaymentContainer = styled.div`
 `
 
 export const StyledPaymentForm = styled.div`
-    background-color: white;
-    border: 2px solid ${COLORS.white};
-    border-radius: 1rem;
-    box-shadow: ${props => props.boxShadow || `.25rem .25rem .75rem .5rem ${COLORS.boxShadowGray}`};
-    display: flex;
-    flex-flow: column wrap;
-    margin: 2rem auto;
-    padding: 1rem;
-    pointer-events: ${props => props.pointerEvents || "all" };
-    position: relative;
-    h2 {
-      padding: 1rem 0;
-      text-align: center;
-    }
-    .digitalWallet {
-      padding: .5rem;
-    }
+  background-color: white;
+  border: 2px solid ${COLORS.white};
+  border-radius: 1rem;
+  box-shadow: ${props => props.boxShadow || `.25rem .25rem .75rem .5rem ${COLORS.boxShadowGray}`};
+  display: flex;
+  flex-flow: column wrap;
+  margin: 2rem auto;
+  padding: 1rem;
+  pointer-events: ${props => props.pointerEvents || "all" };
+  position: relative;
+  h2 {
+    padding: 1rem 0;
+    text-align: center;
+    
+  .digitalWallet {
+    padding: .5rem;
+  }
 `
