@@ -1,12 +1,12 @@
-const CONTACT_API_HOST = process.env.CONTACT_API_HOST
-const CONTACT_API_PROTOCOL = process.env.CONTACT_API_PROTOCOL
+const CONTACT_API_HOST = process.env.CONTACT_API_HOST || 'contact-api'
+const CONTACT_API_PROTOCOL = process.env.CONTACT_API_PROTOCOL || 'http'
 const CONTACT_API_PORT = process.env.CONTACT_API_SERVICE_PORT || process.env.CONTACT_API_PORT
-const MENU_API_HOST = process.env.MENU_API_HOST
-const MENU_API_PROTOCOL = process.env.MENU_API_PROTOCOL
+const MENU_API_HOST = process.env.MENU_API_HOST || 'menu-api'
+const MENU_API_PROTOCOL = process.env.MENU_API_PROTOCOL || 'http'
 const MENU_API_PORT = process.env.MENU_API_SERVICE_PORT || process.env.MENU_API_PORT
-const MERCH_API_HOST = process.env.MERCH_API_HOST
+const MERCH_API_HOST = process.env.MERCH_API_HOST || 'merch-api'
 const MERCH_API_PROTOCOL = process.env.MERCH_API_PROTOCOL || 'http'
-const MERCH_API_PORT = process.env.MERCH_API_SERVICE_PORT
+const MERCH_API_PORT = process.env.MERCH_API_SERVICE_PORT || process.env.MERCH_API_PORT
 
 function makeUrl(port, protocol, host) {
   const url = port ? `${protocol}://${host}:${port}` : `${protocol}://${host}`
